requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
black==24.10.0
    # via -r requirements/dev.in
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via black
coloredlogs==15.0.1
    # via onnxruntime-gpu
distlib==0.3.9
    # via virtualenv
exceptiongroup==1.2.2
    # via pytest
filelock==3.16.1
    # via
    #   torch
    #   triton
    #   virtualenv
flake8==7.1.1
    # via -r requirements/dev.in
flatbuffers==24.3.25
    # via onnxruntime-gpu
fsspec==2024.10.0
    # via torch
humanfriendly==10.0
    # via coloredlogs
identify==2.6.3
    # via pre-commit
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/dev.in
jinja2==3.1.4
    # via torch
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
ndi-python==5.1.1.2
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
networkx==3.4.2
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==2.1.3
    # via
    #   -r /home/geri/work/tools/ndi_recording/requirements/prod.in
    #   ndi-python
    #   onnxruntime-gpu
    #   opencv-python
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
onnxruntime-gpu==1.20.1
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
opencv-python==4.10.0.84
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
packaging==24.2
    # via
    #   black
    #   onnxruntime-gpu
    #   pytest
pathspec==0.12.1
    # via black
pillow==11.0.0
    # via
    #   -r /home/geri/work/tools/ndi_recording/requirements/prod.in
    #   torchvision
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements/dev.in
protobuf==5.29.1
    # via onnxruntime-gpu
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pytest==8.3.4
    # via -r requirements/dev.in
pyyaml==6.0.2
    # via pre-commit
sympy==1.13.1
    # via
    #   onnxruntime-gpu
    #   torch
tomli==2.2.1
    # via
    #   black
    #   pytest
torch==2.5.1
    # via torchvision
torchvision==0.20.1
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
tqdm==4.67.1
    # via -r /home/geri/work/tools/ndi_recording/requirements/prod.in
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   black
    #   torch
virtualenv==20.28.0
    # via pre-commit
